+++ Funcky.Test/FunctionalClass/NoOperationTest.cs
@@ -8,7 +8,7 @@ namespace Funcky.Test.FunctionalClass
             var none = Option<int>.None;
 
             var sideEffect = 0;
-            none.Match(none: Functional.NoOperation, some: i => sideEffect = i);
+            none.Switch(none: Functional.NoOperation, some: i => sideEffect = i);
 
             Assert.Equal(0, sideEffect);
         }