+++ Funcky/DataTypes/EitherOrBoth.cs
@@ -75,7 +75,7 @@ namespace Funcky.DataTypes
             };
 
         [UseWithArgumentNames]
-        public void Match(Action<TLeft> left, Action<TRight> right, Action<TLeft, TRight> both)
+        public void Switch(Action<TLeft> left, Action<TRight> right, Action<TLeft, TRight> both)
         {
             switch (_side)
             {