+++ changelog.md
@@ -6,6 +6,7 @@ Funcky adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Removed in 2.x deprecated APIs 
 * Moved retry policies to a better namespace
 * Moved all Async Code to funcky-async
+* Use Switch instead of Match for Actions to improve type inference
 
 ## Funcky 2.7.0 | Funcky.XUnit 1.0.0 | Funcky.Analyzers 1.1.0
 This release is the last non-breaking release for Funcky before 3.0.